requirements.txt
# clrs benchmark dependecies
absl-py>=0.13.0
attrs>=21.4.0
chex>=0.0.8
dm-haiku>=0.0.4
jax>=0.2.18
jaxlib>=0.1.69
numpy>=1.21.1
opt-einsum>=3.3.0
optax>=0.0.9
six>=1.16.0
tensorflow>=2.9.0
tfds-nightly==4.5.2.dev202204190046
toolz>=0.11.1
torch>=2.0.0
torchvision>=0.15.1
transformers>=4.28.0
dm-clrs @ git+git://github.com/deepmind/clrs.git
# for gpu support https://github.com/deepmind/clrs/issues/98

# images
pillow


# common ml packages
torch>=2.0.0
torchvision
transformers
scikit-learn
xgboost
keras
scipy
numpy
pandas
matplotlib
seaborn
tqdm
kaggle
# tensorrt

# llm
openai
anthropic
langchain
sentencepiece
spacy
nltk
sacrebleu
rouge-score
peft
accelerate
evaluate

# graph
torch-geometric==2.3.1
ogb
networkx


#others
cchardet
pytest
dacite

#open
google-generativeai